requirements.txt
python-telegram-bot==20.3
openai==1.3.5
Flask==2.2.5
gunicorn==21.2.0
